pyproject.toml
[project]
name = "churn-prediction-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jupyter>=1.1.1",
    "kaggle>=1.7.4.5",
    "matplotlib>=3.10.3",
    "numpy>=2.3.0",
    "pandas>=2.3.0",
    "plotly>=6.1.2",
    "polars>=1.31.0",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "seaborn>=0.13.2",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "mypy>=1.16.1",
    "pytest>=8.4.1",
    "ruff>=0.12.0",
]
